refactor python code
