+++ jupyter/data/configs/classification-2d.json
@@ -1,0 +1,17 @@
+{
+    "job_id": "classification-2d",
+    "rows": 100,
+    "cols": 3,
+    "memory_limit": 50000000,
+    "threads": 3,
+    "results_field": "ml",
+      "categorical_fields": ["target"],
+    "analysis": {
+      "name": "classification",
+      "parameters": {
+        "dependent_variable": "target",
+        "num_classes": 2,
+        "class_assignment_objective": "maximize_accuracy"
+      }
+    }
+  }
