+++ jupyter/data/configs/regression-2d.json
@@ -1,0 +1,15 @@
+{
+    "job_id": "regression-2d",
+    "rows": 100,
+    "cols": 3,
+    "memory_limit": 50000000,
+    "threads": 3,
+    "results_field": "ml",
+    "analysis": {
+      "name": "regression",
+      "parameters": {
+        "dependent_variable": "target",
+        
+      }
+    }
+  }
