+++ jupyter/requirements.txt
@@ -11,3 +11,4 @@ graphviz
 pathlib2            ==2.3.5
 nbsmoke             ==0.5.0
 bump2version        ==1.0.1
+scikit-multiflow    ==0.5.3
